fix(ui): Adjust scroll threshold and add mobile animation
Update the scroll threshold for the floating action bar in `SetDetailView` to 350px.
Also, enhance the mobile floating bar animation in `SetDetailView` with a scale-in effect for a more engaging user experience.
Additionally, minor UI adjustments were made to the search input styling in `Header.tsx`.
Introduced a check for `GOOGLE_CLIENT_ID` in `Login.tsx` for improved robustness.
Added `ChevronRight` icon to `CreateSet.tsx`.
Replaced `AlertTriangle` with `TriangleAlert` in `Login.tsx` for consistency.
Added `Loader2` icon to `Login.tsx`.